# adapter/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Route Service API
  version: 1.0.0
  description: |
    Route planning: brouter proxy (semaphore-protected), draft persistence,
    elevation profile (Copernicus DEM). Plus admin endpoints to manage BRouter
    segments (.rd5) and profiles.
servers:
  - url: http://localhost:8085
    description: Local development server
paths:
...
  /admin/route/segments:
    get:
      summary: List installed + available segments (filtered to Europe by default)
      operationId: listSegments
      tags: [Segments]
      parameters:
        - name: includeWorld
          in: query
          required: false
          description: If true, return all globe segments. Default false = Europe only.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Catalog of all known segments + flags which are installed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SegmentResponse'
  /admin/route/segments/download:
    post:
      summary: Start async download of a segment
      operationId: downloadSegment
      tags: [Segments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadSegmentRequest'
      responses:
        '202':
          description: Task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SegmentTaskResponse'
  /admin/route/segments/{name}:
    delete:
      summary: Delete an installed segment (.rd5)
      operationId: deleteSegment
      tags: [Segments]
      parameters:
        - name: name
          in: path
          required: true
          description: Segment name without .rd5 (e.g. "E20_N50")
          schema:
            type: string
      responses:
        '204':
          description: Deleted (workers will be rolling-restarted)
...
  /admin/route/logs/{name}:
    get:
      summary: Tail the last N lines of a log file (plain text)
      operationId: tailLog
      tags: [Status]
      parameters:
        - name: name
          in: path
          required: true
          description: Log source name from /admin/route/logs (e.g. worker-0, nginx-access)
          schema:
            type: string
        - name: lines
          in: query
          required: false
          schema:
            type: integer
            default: 200
            minimum: 1
            maximum: 5000
      responses:
        '200':
          description: Tail content
          content:
            text/plain:
              schema:
                type: string
  /admin/route/workers/{index}/restart:
    post:
      summary: Restart a single BRouter worker
      operationId: restartWorker
      tags: [Status]
      parameters:
        - name: index
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '204':
          description: Triggered
...
components:
  schemas:
...
    SegmentResponse:
      type: object
      required: [name, lonStart, latStart, installed]
      properties:
        name:
          type: string
          description: e.g. "E20_N50" (no .rd5 suffix)
          example: E20_N50
        lonStart:
          type: integer
          description: West longitude of 5°×5° tile (signed)
          example: 20
        latStart:
          type: integer
          description: South latitude of 5°×5° tile (signed)
          example: 50
        remoteSizeBytes:
          type: integer
          format: int64
          nullable: true
          description: Size from brouter.de (null if remote listing failed)
        installed:
          type: boolean
        installedSizeBytes:
          type: integer
          format: int64
          nullable: true
        installedAt:
          type: string
          format: date-time
          nullable: true
        outdated:
          type: boolean
          nullable: true
          description: true if installed file size differs from remote
    DownloadSegmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Segment name without .rd5 suffix
          example: E20_N50
    SegmentTaskResponse:
      type: object
      required: [taskId, segmentName, status]
      properties:
        taskId:
          type: string
        segmentName:
          type: string
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED]
...